Update testing for e2e test in single py script
